Fix date comparison

+++ partner_multi_relation/models/res_partner_relation_type.py
@@ -152,7 +152,8 @@ class ResPartnerRelationType(models.Model):
                     # already ended:
                     cutoff_date = fields.Date.today()
                     for relation in invalid_relations:
-                        if relation.date_start >= cutoff_date:
+                        if (relation.date_start and
+                                relation.date_start >= cutoff_date):
                             relation.unlink()
                         elif (not relation.date_end or
                                 relation.date_end > cutoff_date):
